requirements.txt
pandas
requests
mistralai
openpyxl
psycopg2-binary
tqdm
apache-airflow